extract param change handler
add dependency to react-datepicker
